[XPD] Migrate to SettingsBasePreferenceFragment
The motivation for this CL is that this is a required step (in
addition to enabling the feature flag) for enabling expressive design.

The required changes are generally for all fragments to inherit from
SettingsBasePreferenceFragment instead of PreferenceFragmentCompat.

We also apply theme the "Theme_SubSettingsBase_Expressive" (when
expressive design is enabled) to all of our Settings pages.

The intended scope of this CL is handheld fragments and excludes
Safety Center.

Bug: 375480013
Test: atest CtsPermissionUiTestCases
Flag: com.android.settingslib.widget.theme.flags.is_expressive_design_enabled
Relnote: N/A
Change-Id: I829936db39fc7fc5d78c9db2f00ae5d0b4dd0c32