Run AppPermissionsTest from Android 16+
ag/31382100 (which adds the CREATE_VIRTUAL_DEVICE permission to the com.android.shell package) isn't there in Android 15, but its there in Android 16. So, AppPermissionsTest fails on 25Q1 builds, as it relies on shell for the above permission. So, we annotate the test to run on Android 16+, similar to DeviceAwarePermissionGrantTest.

Relnote: Annotate AppPermissionsTest to run on Android 16 and above
Test: atest AppPermissionsTest
Fixes: 406924219
LOW_COVERAGE_REASON=TEST_ONLY
Flag: EXEMPT test fix
Change-Id: I5ff1a33d146112c048a18adbbe2753b76d55078f