Use SettingsLib titleless preference category in SC.
This replaces usages of our custom no-title layout category. The updated
preference comparison code is required since this
SettingsTitlelessPreferenceCategory isn't open, and is written in a way
where we could replace our custom ComparablePreferenceCategory with
vanilla PreferenceCategory usages with no change in functionality. But
that change is left to a later commit to minimize the diff here.

Fixes: 406934282
Test: atest SafetyCenterActivityFunctionalTestCases SafetyCenterSubpagesTestCases and manual
Relnote: UI fix for expressive design
Flag: EXEMPT bugfix
Change-Id: I732fceb4290ece5b7eb6647d4558ae7b54eccee0