Store current and past certificate digests in a LinkedHashSet
Permissions are backed up to and restored from an XML document. We
discovered that the XML output that is produced by the BackupHelper is
non-deterministic, because it iterates over HashSets. The
non-deterministic output means we have to upload this document ~every
time and don't benefit from incrementally updating backups. This is
causing unnecessary network traffic and storage use for the backup
service.

Test: Manually compared output of repeated backup attempts
Bug: 372237049
Flag: EXEMPT bugfix
Relnote: bug fix
Change-Id: I662d00655d6e3fa697cbedaad2789fd8d9c4163b