Add config_enableExpressiveDesignInSafetyCenter
This creates a boolean resource that must be set to true for expressive
design to be used in Safety Center. It's overlayable for OEMs to opt out
of expressive design.

The boolean is currently unused since Safety Center completely disabled
expressive design for now, but will be used once expressive design
support is re-added. Adding it now so the overlayable is present in the
prebuilt for the 25Q3 release.

Bug: 414542006
Test: treehugger
Flag: com.android.settingslib.widget.theme.flags.is_expressive_design_enabled
Relnote: Adds overlayable prop for opting out of expressive design in
Safety Center

Change-Id: I6b817308034fac16129afc352832938b0916c850