Preserve non-clipping behavior for preference category
The entire view hierarchy was non-clipping on V since ag/28195678, but
that was broken by ag/31875944 in B and it broke V OEM customization on
preference category that relied on it.

As an urgent fix for M-2025-05, we can simply set the two layouts for
preference category to be non-clipping again, while limiting any other
potential changes (even if that may fix something else).

This only affects the category headings in permission settings pages,
and should be a no-op for devices without customization since we don't
have anything to be clipped in preference category in AOSP.

LOW_COVERAGE_REASON=NON_CODE_ONLY

Bug: 412863862
Flag: EXEMPT bugfix
Relnote: N/A
Change-Id: Ie091289cebd5a3983d8d5c30c2c5b85c254ba7ef