Add title and description for recommended assistants
All of config_recommendedAssistantsBy,
config_recommendedAssistantsDescription and config_recommendedAssistants
need to be overlaid for the recommendation to be enabled.

FooterPreference is used instead of TopIntroPreference even for
non-footer description in order to maintain compatibility with previous
platforms and reduce risk.

Bug: 388234667
Flag: com.android.permission.flags.default_apps_recommendation_enabled
Test: manual with ag/31585104
Relnote: N/A
Change-Id: I3afb4ca99738b163b27c31f1781d4cea97bb82bf