Turn off SC expressive config flag by default.
We'll likely enable this in the future with a Google/Pixel specific
overlay, but leaving a note that it might change just in case.

This config value is still not read in SC code, which hardcodes
expressive design to be off.

Bug: 414542006
Test: treehugger
Flag: com.android.settingslib.widget.theme.flags.is_expressive_design_enabled
Relnote: Modifies overlayable prop for opting out of expressive design in
Safety Center

Change-Id: I2730215ea5cc5fb3961ca9f17b6e9da77365a2e8