Add deep link support for device state in PermissionController
ManagePermissionsActivity is protected by GRANT_RUNTIME_PERMISSION which the Gemini app doesn't have. So we need to use the intermediate ManagePermissionsActivityTrampoline in a similar way of providing deeplink in the Settings search results, which relies on the password key in the Intent for verification.

Add new deeplinks for the following pages:
- security & privacy > privacy > permission manager > additional permissions
- apps > {packageName} > permissions

Bug: 414364331
Test:  adb shell command to start the activity and verify the page is displayed on the phone, e.g. adb shell am start -a com.android.permissioncontroller.devicestate.action.MANAGE_APP_PERMISSIONS -e "com.android.permissioncontroller.devicestate.key" "25a505a5-e92f-449f-9a85-7d29987e3364" -e "android.intent.extra.PACKAGE_NAME" "com.google.android.gm"
Flag: EXEMPT BUG_FIX
Relnote: N/A
Change-Id: Ia7bf60c3ecefb2c9dc4b4a4363cf47cd672f72cc
LOW_COVERAGE_REASON=b/411150350