Make debug logs check isLoggable
Replace DEBUG_HIBERNATION_POLICY with an isLoggable check so that it can
be turned on in CTS tests for debuggability. Also add a log for exempt
services for the HibernationPolicy tag.

Also clean up some dead code for DEBUG_OVERRIDE_THRESHOLDS which is
always "false".

Bug: 419660996
Test: atest CtsHibernationTestCases
Flag: EXEMPT non-functional change
Relnote: N/A
Change-Id: I4fbf5541b8800477395649bbd3291bae5f52a5d5