Log when deviceId is null in a flaky test
This is a test with low flakiness rate. At this point, we don't know when the
test fails whether it's due to deviceId passed in is null or uid doesn't
exist in mUidDeviceIdsMap. So add a logging to give us more clarity.

Bug: 421987526
Test: atest android.permission.cts.PermissionUpdateListenerTest#testDefaultDeviceGrantPermissionNotifyListener
Flag: EXEMPT TEST_ONLY
Relnote: N/A
Change-Id: I24ebef4148981224473e1c3109e2f4dd61ac5f0d
LOW_COVERAGE_REASON=TEST_ONLY