Preserve non-clipping behavior for legacy app permission switch parent
The entire view hierarchy was non-clipping on V since ag/28195678, but
that was broken by ag/31875944 in B and it broke V OEM switch appearance
that relied on it.

As an urgent fix for Mainline, we can simply set the parent layout to be
non-clipping again, while limiting any other potential changes (either
good or bad).

This only affects this particular layout, and should be a no-op for
devices without customization since we don't have anything to be clipped
inside it in AOSP.

This is similar to ag/33358288 for b/412863862.

LOW_COVERAGE_REASON=NON_CODE_ONLY

Bug: 427305183
Flag: EXEMPT bugfix
Relnote: N/A
Change-Id: Id70121fd7eec797e5722f94a0e7ab5b155649ee2