Fix Privacy dashboard page window insets
Somehow we have never enabled edge to edge for Permission manager and
Privacy dashboard pages. It's fine for Permission manager pages because
that is using navigation graph which has a base layout that wrapped the
CoordinatorLayout with a ConstraintLayout that applied the window
insets. However that isn't happening for Privacy dashboard so it still
needs the edge to edge hack from SettingsLib, which is basically
applying window insets on root layout (android.id.content) and then
making child views non-clipping.

Fixes: 412269399
Test: manually confirm e2e working on privacy dashboard, permission
manager and data sharing updates.
Flag: EXEMPT bug fix
Relnote: N/A
Change-Id: I78eb4fef56361a86a5aac8b4d05d93765ab44291