Add agents privacy dashboard
Add support for showing agent activity in privacy dashboard.

This includes adding needed methods in the repository and adding new UseCases:
AppFunctionRepository
GetAppFunctionAgentUsageUseCase

Also changes the existing Privacy Dashboard Fragment and ViewModel to support showing agent activity category. The agent activity category is gated with the flag (mentioned in the commit footer). The agent activity entries can be filtered with the 24h/7d toggle.

However, the show system is not yet supported. In this CL, it will show all entries no matter which of show/hide system is toggled.

In the new design, we also change the background of the app permission expand button. In the past, we've been using the expand button that comes with the PreferenceGroupAdapter. However, this limits our flexibility to customize button (e.g. changing the style or the behavior) and leads us to be using hacky ways to customize the title and summary (we need to manually find the preference category that contains the button then get the preference with order 999). Hence, we changed to be using ExpandButtonPreference. This also helps us be the align with the BC25 styles and widgets.

Bug: 436261702
Flag: com.android.permissioncontroller.flags.privacy_dashboard_agent_activity_enabled
Low-Coverage-Reason: FLAG_NOT_ENABLED
Test: PermissionUsageViewModelTest and GetAppFunctionAgentUsageUseCaseTest
Relnote: Add agents privacy dashboard
Change-Id: Iceccc277241eddf5a320bb64c76a1868f337aa96