Fix BIND_ON_DEVICE_INTELLIGENCE_SERVICE to be signature only permission
BIND_ON_DEVICE_INTELLIGENCE_SERVICE was incorrectly marked as signature|privileged when infact it was meant to be signature only i.e. only system can be granted this permission. Although this is not a security issue currently, this is not intended, is incorrect and is leading to confusion.

There are no active usages of this in privileged apps - so it should be safe to remove it.

Bug:464230790
Flag: EXEMPT BUGFIX
Relnote: Fix BIND_ON_DEVICE_INTELLIGENCE_SERVICE to be signature only permission
Change-Id: I6d73462b75a4871100e440147b097eef22c7b5b2