Change GET_ATTRIBUTIONS to GET_ATTRIBUTIONS_LONG for U+
PermissionsApps uses the flag GET_ATTRIBUTIONS when it calls into
various PackageManager methods. This can can cause an unintendedn sign
extension, and it seems to mistakenly act as if the GET_APP_LOCK_INFO
flag was added, which results in a SecurityException for all apps
without the LOCK_APPS permission. Change this to GET_ATTRIBUTIONS_LONG
for U+ so that the long flags don't get accidentally filled out, and the
exception is not thrown.

Flag: EXEMPT BUGFIX
Relnote: bug fix
Low-Coverage-Reason: Test flakiness
Test: atest PermissionControllerHostTest
Fixes: 465847514
Change-Id: I756c5813538c9f91cc33d415858b723cdbb6ee3a